Iterator: Tracking maximum value
Added findLargest operation to Node
During insertion, after the Node is found, the node is compared to
the current maximum-containing-node. If they are the same, then
we search for and update the maximum after the insertion operation.

The search is completed in constant time because there are only two
possible nodes it may be in - the same one or the right neighbour

Also added operator+ for the iterator.