Iterator (#2)
* Basic Iterator Operations

 * Wrote test and implementation for iterating forward in trivial cases
 * Replaced iterator class with a typedef to const_iterator because it must be const
 * Removed duplicate typedefs for const_reference and const_pointers in the const_iterator

* Minor renaming, iterate test improvements

* Added ostream operator<< overload for testing

* Iterator: Tracking maximum value

Added findLargest operation to Node
During insertion, after the Node is found, the node is compared to
the current maximum-containing-node. If they are the same, then
we search for and update the maximum after the insertion operation.

The search is completed in constant time because there are only two
possible nodes it may be in - the same one or the right neighbour

Also added operator+ for the iterator.

* Removed after_end and after_rend iterators

The after_end class fields can be generated from the begin and rbegin iterators
when they are required. No need to keep an instance alive for each class.
Also updated the rbegin iterator and added accessor functions. However,
the operator-- hasn't been implemented yet so they aren't particularly useful.

* Print time taken for all tests to run

* Complete working iterator and reverse iterator

Added the operator-- for the iterator and generated the correct reverse_iterator
Modified some test cases to also check for reverse_iterator correctness

* Moved iterator implementation to separate file

Created iter_operations.ipp and moved iterator implementation there
Minor renaming
Added some more test cases