Implemented remove with shrink, remove on internal node
Remove and shrink occurs when the parent and the sibling don't
have enough keys for a transfer. The three nodes are fused into one.

Removing internal nodes is logically straightforward:
the key is swapped with its predecessor (which is a leaf node),
then the key is removed using the previously implemented functionality