Erase - Major refactoring
The previous implementation of erasing had a fatal flaw: In cases where an
underflow could not be resolved, the underflow must propagate upwards until it
is handled. To solve this, a proper implementation of erase should resolve
possible underflows while it is traversing down the tree in search of the value
to be erased.

The same logic is still used (cw/ccw rotations are now transferFrom[Right|Left],
and the fusion and shrink operations behave similarly. However, these are now made
generic enough to work for internal nodes, meaning that children are also handled now.

There is still room for improvement and optimizations for when a full leaf node is going
to be removed