Erasing (#5)
* Fixed Node const correctness

There are still some instances where TwoFourTree has non-const access to
internal Nodes. The implementation needs to be changed to fix this,
which was left for another time

* Debug: Added additional case to validate

It now also checks the number of children and number of keys.
In a TwoFourTree, for n as the number of keys in a node, every internal
node must have n+1 children.

* Erase

A proper implementation of erase should resolve possible underflows while it is traversing 
down the tree in search of the value to be erased.

The same logic is still used (cw/ccw rotations are now transferFrom[Right|Left],
and the fusion and shrink operations behave similarly. However, these are now made
generic enough to work for internal nodes, meaning that children are also handled now.

There is still room for improvement and optimizations for when a full leaf node is going
to be removed

* Changed the traversals to use traverse_step

* Removed unused function signatures, minor optimizations and fixes

* Handled erasing missing element

* Added debug capabilities

The validate function now verifies that each node has the correct number
of children, and that the sorting of the children is consistent

* Updated begin/end iterators on erase

Temporary crude solution is to search the entire tree for the iterators
on every single erasure

* Added iterator test case after erasing values

* Created test class that may be used as the type of object contained by the Tree